+++ lib/middleware/index.ts
@@ -2,6 +2,8 @@ export { middleware } from "./middleware";
 export { cors } from "./cors";
 export { sessionExists } from "./sessionExists";
 export { validBearerToken } from "./validBearerToken";
+export { validFormSubmission } from "./validFormSubmission";
+export { validQueryParams } from "./validQueryParams";
 export { validTemporaryToken } from "./validTemporaryToken";
 export { jsonValidator } from "./jsonValidator";
 export { csrfProtected } from "./csrfProtected";